MCN/requirements.txt
torch
torchvision
torchaudio
numpy
opencv-python
spacy
tensorboard
tensorboardX
albumentations
pillow
pycocotools
pyyaml
omegaconf==2.0.6
cloudpickle
hydra-core==1.0.7
termcolor
